as per updation noted

+++ schema.sql
@@ -28,8 +28,8 @@ FOREIGN KEY (uid) REFERENCES user_detail(uid)
 CREATE TABLE savepolicy(
 uid int,
 ID int NOT NULL AUTO_INCREMENT,
-consignmentWeight int NOT NULL,
-consignmentValue int NOT NULL,
+consignmentWeight BIGINT(100) NOT NULL,
+consignmentValue BIGINT(100) NOT NULL,
 contractType VARCHAR(100) NOT NULL,
 policyType VARCHAR(100) NOT NULL,
 PRIMARY KEY(ID),
